If parsing, fixed while parsing, changed error messages
Also deleted EOL (EOF) check from END case from start(). Had to do it because it made impossible for me to check if 'else' was ended with 'end'. It could've been ended by 'else' instead, so I decided to copy the code from start() to the functions, which have to be ended by 'end' keyword. I think it's more fancy to do it this way, instead of creating new variable in pData structure to count 'else' keywords.